Renamed productCreate file
